+++ src/components/QueuePanel.tsx
@@ -9,7 +9,7 @@ import { useOrbitStore } from '../store/orbitStore';
 import OrbitGuestQueue from './OrbitGuestQueue';
 import OrbitQueueHead from './OrbitQueueHead';
 import HostApprovalQueue from './HostApprovalQueue';
-import { Play, Trash2, Save, FolderOpen, Shuffle, Infinity, Waves, MicVocal, ListMusic, Check, MoveRight, Radio, Info, Share2 } from 'lucide-react';
+import { Play, MicVocal, ListMusic, Radio, Info } from 'lucide-react';
 import { usePlaylistStore } from '../store/playlistStore';
 import { useCachedUrl } from './CachedImage';
 import { useTranslation } from 'react-i18next';
@@ -25,7 +25,7 @@ import NowPlayingInfo from './NowPlayingInfo';
 import { TFunction } from 'i18next';
 import OverlayScrollArea from './OverlayScrollArea';
 import { useLuckyMixStore } from '../store/luckyMixStore';
-import { useQueueToolbarStore, QueueToolbarButtonId } from '../store/queueToolbarStore';
+import { useQueueToolbarStore } from '../store/queueToolbarStore';
 import {
   DurationMode,
   formatTime,
@@ -36,6 +36,7 @@ import { QueueHeader } from './queuePanel/QueueHeader';
 import { QueueCurrentTrack } from './queuePanel/QueueCurrentTrack';
 import { useQueuePanelDrag } from '../hooks/useQueuePanelDrag';
 import { useQueueLufsTgtPopover } from '../hooks/useQueueLufsTgtPopover';
+import { QueueToolbar } from './queuePanel/QueueToolbar';
 
 export default function QueuePanel() {
   const orbitRole = useOrbitStore(s => s.role);
@@ -131,11 +132,8 @@ function QueuePanelHostOrSolo() {
   const setIsNowPlayingCollapsed = useAuthStore(s => s.setQueueNowPlayingCollapsed);
   const toolbarButtons = useQueueToolbarStore(s => s.buttons);
   const [durationMode, setDurationMode] = useState<DurationMode>('total');
-  const [showCrossfadePopover, setShowCrossfadePopover] = useState(false);
   const expandReplayGain = useThemeStore(s => s.expandReplayGain);
   const setExpandReplayGain = useThemeStore(s => s.setExpandReplayGain);
-  const crossfadeBtnRef = useRef<HTMLButtonElement>(null);
-  const crossfadePopoverRef = useRef<HTMLDivElement>(null);
   const reanalyzeLoudnessForTrack = usePlayerStore(s => s.reanalyzeLoudnessForTrack);
   const authLoudnessTargetLufs = useAuthStore(s => s.loudnessTargetLufs);
   const setLoudnessTargetLufs = useAuthStore(s => s.setLoudnessTargetLufs);
@@ -148,19 +146,6 @@ function QueuePanelHostOrSolo() {
     lufsTgtBtnRef,
     lufsTgtMenuRef,
   } = useQueueLufsTgtPopover(expandReplayGain);
-
-  useEffect(() => {
-    if (!showCrossfadePopover) return;
-    const handle = (e: MouseEvent) => {
-      if (
-        crossfadeBtnRef.current?.contains(e.target as Node) ||
-        crossfadePopoverRef.current?.contains(e.target as Node)
-      ) return;
-      setShowCrossfadePopover(false);
-    };
-    document.addEventListener('mousedown', handle);
-    return () => document.removeEventListener('mousedown', handle);
-  }, [showCrossfadePopover]);
 
   const queueListRef = useRef<HTMLDivElement>(null);
 
@@ -331,132 +316,26 @@ function QueuePanelHostOrSolo() {
 
       {activeTab === 'queue' ? (<>
         {!isNowPlayingCollapsed && toolbarButtons.some(b => b.visible && b.id !== 'separator') && (
-          <div className="queue-toolbar">
-            {toolbarButtons.map((btn, idx) => {
-              if (!btn.visible) return null;
-
-              switch (btn.id as QueueToolbarButtonId) {
-                case 'shuffle':
-                  return (
-                    <button key={btn.id} className="queue-round-btn" onClick={() => shuffleQueue()} disabled={queue.length < 2} data-tooltip={t('queue.shuffle')} aria-label={t('queue.shuffle')}>
-                      <Shuffle size={13} />
-                    </button>
-                  );
-                case 'save':
-                  return (
-                    <button
-                      key={btn.id}
-                      className={`queue-round-btn${saveState === 'saved' ? ' active' : ''}`}
-                      onClick={handleSave}
-                      disabled={saveState === 'saving'}
-                      data-tooltip={activePlaylist ? `${t('queue.updatePlaylist')}: ${activePlaylist.name}` : t('queue.savePlaylist')}
-                      aria-label={t('queue.savePlaylist')}
-                    >
-                      {saveState === 'saved' ? <Check size={13} /> : <Save size={13} />}
-                    </button>
-                  );
-                case 'load':
-                  return (
-                    <button key={btn.id} className="queue-round-btn" onClick={handleLoad} data-tooltip={t('queue.loadPlaylist')} aria-label={t('queue.loadPlaylist')}>
-                      <FolderOpen size={13} />
-                    </button>
-                  );
-                case 'share':
-                  return (
-                    <button
-                      key={btn.id}
-                      className="queue-round-btn"
-                      onClick={() => void handleCopyQueueShare()}
-                      data-tooltip={t('queue.shareQueue')}
-                      aria-label={t('queue.shareQueue')}
-                    >
-                      <Share2 size={13} />
-                    </button>
-                  );
-                case 'clear':
-                  return (
-                    <button key={btn.id} className="queue-round-btn" onClick={handleClear} data-tooltip={t('queue.clear')} aria-label={t('queue.clear')}>
-                      <Trash2 size={13} />
-                    </button>
-                  );
-                case 'separator':
-                  return <div key={btn.id} className="queue-toolbar-sep" />;
-                case 'gapless':
-                  return (
-                    <button
-                      key={btn.id}
-                      className={`queue-round-btn${gaplessEnabled ? ' active' : ''}`}
-                      onClick={() => { setCrossfadeEnabled(false); setShowCrossfadePopover(false); setGaplessEnabled(!gaplessEnabled); }}
-                      data-tooltip={t('queue.gapless')}
-                      aria-label={t('queue.gapless')}
-                    >
-                      <MoveRight size={13} />
-                    </button>
-                  );
-                case 'crossfade':
-                  return (
-                    <div key={btn.id} style={{ position: 'relative' }}>
-                      <button
-                        ref={crossfadeBtnRef}
-                        className={`queue-round-btn${crossfadeEnabled || showCrossfadePopover ? ' active' : ''}`}
-                        onClick={() => {
-                          if (crossfadeEnabled) {
-                            setCrossfadeEnabled(false);
-                            setShowCrossfadePopover(false);
-                          } else {
-                            setGaplessEnabled(false);
-                            setCrossfadeEnabled(true);
-                            setShowCrossfadePopover(true);
-                          }
-                        }}
-                        data-tooltip={showCrossfadePopover ? undefined : t('queue.crossfade')}
-                        aria-label={t('queue.crossfade')}
-                      >
-                        <Waves size={13} />
-                      </button>
-                      {showCrossfadePopover && (
-                        <div className="crossfade-popover" ref={crossfadePopoverRef}>
-                          <div className="crossfade-popover-label">
-                            <Waves size={11} />
-                            {t('queue.crossfade')}
-                            <span className="crossfade-popover-value">{crossfadeSecs.toFixed(1)} s</span>
-                          </div>
-                          <input
-                            type="range"
-                            min={0.1}
-                            max={10}
-                            step={0.1}
-                            value={crossfadeSecs}
-                            onChange={e => {
-                              setCrossfadeSecs(parseFloat(e.target.value));
-                              setCrossfadeEnabled(true);
-                            }}
-                            className="crossfade-popover-slider"
-                          />
-                          <div className="crossfade-popover-range">
-                            <span>0.1s</span><span>10s</span>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  );
-                case 'infinite':
-                  return (
-                    <button
-                      key={btn.id}
-                      className={`queue-round-btn${infiniteQueueEnabled ? ' active' : ''}`}
-                      onClick={() => setInfiniteQueueEnabled(!infiniteQueueEnabled)}
-                      data-tooltip={t('queue.infiniteQueue')}
-                      aria-label={t('queue.infiniteQueue')}
-                    >
-                      <Infinity size={13} />
-                    </button>
-                  );
-                default:
-                  return null;
-              }
-            })}
-          </div>
+          <QueueToolbar
+            queue={queue}
+            activePlaylist={activePlaylist}
+            saveState={saveState}
+            toolbarButtons={toolbarButtons}
+            shuffleQueue={shuffleQueue}
+            handleSave={handleSave}
+            handleLoad={handleLoad}
+            handleCopyQueueShare={handleCopyQueueShare}
+            handleClear={handleClear}
+            gaplessEnabled={gaplessEnabled}
+            setGaplessEnabled={setGaplessEnabled}
+            crossfadeEnabled={crossfadeEnabled}
+            setCrossfadeEnabled={setCrossfadeEnabled}
+            crossfadeSecs={crossfadeSecs}
+            setCrossfadeSecs={setCrossfadeSecs}
+            infiniteQueueEnabled={infiniteQueueEnabled}
+            setInfiniteQueueEnabled={setInfiniteQueueEnabled}
+            t={t}
+          />
         )}
 
       {currentTrack && queue.length > 0 && <div className="queue-divider"><span style={{ fontSize: '12px', fontWeight: 600, color: 'var(--text-muted)' }}>{t('queue.nextTracks')}</span></div>}
